requirements.txt
beautifulsoup4==4.10.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
google==3.0.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
idna==3.3
importlib-metadata==4.9.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
